ISSUE #84 * Minor refactor (2).
