+++ minos/common/model/declarative/aggregate/model.py
